refactor: split runtime transport units
