+++ runtime/include/backend/BackendServerRequest.hpp
@@ -1,0 +1,6 @@
+#ifndef GRAVITY_RUNTIME_INCLUDE_BACKEND_BACKENDSERVERREQUEST_HPP_
+#define GRAVITY_RUNTIME_INCLUDE_BACKEND_BACKENDSERVERREQUEST_HPP_
+
+#include "backend/BackendServer.hpp"
+
+#endif // GRAVITY_RUNTIME_INCLUDE_BACKEND_BACKENDSERVERREQUEST_HPP_
