+++ runtime/include/frontend/FrontendRuntimePolling.hpp
@@ -1,0 +1,6 @@
+#ifndef GRAVITY_RUNTIME_INCLUDE_FRONTEND_FRONTENDRUNTIMEPOLLING_HPP_
+#define GRAVITY_RUNTIME_INCLUDE_FRONTEND_FRONTENDRUNTIMEPOLLING_HPP_
+
+#include "frontend/FrontendRuntime.hpp"
+
+#endif // GRAVITY_RUNTIME_INCLUDE_FRONTEND_FRONTENDRUNTIMEPOLLING_HPP_
