+++ runtime/include/protocol/BackendClientIo.hpp
@@ -1,0 +1,6 @@
+#ifndef GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTIO_HPP_
+#define GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTIO_HPP_
+
+#include "protocol/BackendClient.hpp"
+
+#endif // GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTIO_HPP_
