+++ runtime/include/protocol/BackendClientRequests.hpp
@@ -1,0 +1,6 @@
+#ifndef GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTREQUESTS_HPP_
+#define GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTREQUESTS_HPP_
+
+#include "protocol/BackendClient.hpp"
+
+#endif // GRAVITY_RUNTIME_INCLUDE_PROTOCOL_BACKENDCLIENTREQUESTS_HPP_
