+++ runtime/src/protocol/BackendClient.cpp
@@ -1,16 +1,11 @@
 #include "protocol/BackendClient.hpp"
-#include "protocol/BackendProtocol.hpp"
 #include "platform/SocketPlatform.hpp"
 
-#include <algorithm>
-#include <array>
-#include <cctype>
 #include <cstddef>
 #include <exception>
 #include <iostream>
 #include <string>
 #include <string_view>
-#include <utility>
 
 static grav_socket::ConstBytes asBytes(std::string_view text)
 {
@@ -32,11 +27,6 @@ static bool sendAll(grav_socket::Handle socketHandle, grav_socket::ConstBytes by
         offset += static_cast<std::size_t>(sent);
     }
     return true;
-}
-
-static int clampTimeoutMs(int timeoutMs)
-{
-    return grav_socket::clampTimeoutMs(timeoutMs);
 }
 
 static std::string backendClientError(std::string_view operation, std::string_view detail)
@@ -96,24 +86,6 @@ bool BackendClient::connect(const std::string &host, std::uint16_t port)
     }
 }
 
-void BackendClient::setSocketTimeoutMs(int timeoutMs)
-{
-    _socketTimeoutMs = clampTimeoutMs(timeoutMs);
-    if (isConnected()) {
-        grav_socket::setSocketTimeoutMs(_socket, _socketTimeoutMs);
-    }
-}
-
-int BackendClient::socketTimeoutMs() const
-{
-    return _socketTimeoutMs;
-}
-
-void BackendClient::setAuthToken(std::string token)
-{
-    _authToken = std::move(token);
-}
-
 void BackendClient::disconnect()
 {
     try {
@@ -134,65 +106,6 @@ void BackendClient::disconnect()
         _socket = grav_socket::invalidHandle();
         _recvBuffer.clear();
         _networkInitialized = false;
-    }
-}
-
-bool BackendClient::isConnected() const
-{
-    return grav_socket::isValid(_socket);
-}
-
-std::string BackendClient::trim(const std::string &value)
-{
-    const auto begin = std::find_if_not(value.begin(), value.end(), [](unsigned char c) {
-        return std::isspace(c) != 0;
-    });
-    const auto end = std::find_if_not(value.rbegin(), value.rend(), [](unsigned char c) {
-        return std::isspace(c) != 0;
-    }).base();
-    if (begin >= end) {
-        return {};
-    }
-    return std::string(begin, end);
-}
-
-bool BackendClient::readLine(std::string &outLine)
-{
-    try {
-        if (!grav_socket::isValid(_socket)) {
-            return false;
-        }
-
-        std::size_t pos = _recvBuffer.find('\n');
-        while (pos == std::string::npos) {
-            std::array<char, 2048> chunk{};
-            const int received = grav_socket::recvBytes(
-                _socket,
-                grav_socket::MutableBytes{
-                    reinterpret_cast<std::byte *>(chunk.data()),
-                    chunk.size()});
-            if (received <= 0) {
-                return false;
-            }
-            _recvBuffer.append(chunk.data(), static_cast<std::size_t>(received));
-            if (_recvBuffer.size() > (512u * 1024u)) {
-                return false;
-            }
-            pos = _recvBuffer.find('\n');
-        }
-
-        outLine = _recvBuffer.substr(0, pos);
-        _recvBuffer.erase(0, pos + 1);
-        if (!outLine.empty() && outLine.back() == '\r') {
-            outLine.pop_back();
-        }
-        return true;
-    } catch (const std::exception &ex) {
-        std::cerr << backendClientError("readLine", ex.what()) << "\n";
-        return false;
-    } catch (...) {
-        std::cerr << backendClientError("readLine", "non-standard exception") << "\n";
-        return false;
     }
 }
 
@@ -245,87 +158,3 @@ BackendClientResponse BackendClient::sendJson(const std::string &jsonLine)
     }
 }
 
-BackendClientResponse BackendClient::sendCommand(const std::string &cmd, const std::string &fieldsJson)
-{
-    grav_protocol::BackendCommandRequest request{};
-    request.cmd = cmd;
-    request.token = _authToken;
-    return sendJson(grav_protocol::BackendJsonCodec::makeCommandRequest(request, fieldsJson));
-}
-
-BackendClientResponse BackendClient::getStatus(BackendClientStatus &outStatus)
-{
-    try {
-        BackendClientResponse response = sendCommand(std::string(grav_protocol::Status));
-        if (!response.ok) {
-            return response;
-        }
-
-        grav_protocol::BackendStatusPayload parsed{};
-        std::string parseError;
-        if (!grav_protocol::BackendJsonCodec::parseStatusResponse(response.raw, parsed, parseError)) {
-            response.ok = false;
-            response.error = backendClientError("getStatus", parseError);
-            return response;
-        }
-
-        outStatus.steps = parsed.steps;
-        outStatus.dt = parsed.dt;
-        outStatus.paused = parsed.paused;
-        outStatus.faulted = parsed.faulted;
-        outStatus.faultStep = parsed.faultStep;
-        outStatus.faultReason = parsed.faultReason;
-        outStatus.sphEnabled = parsed.sphEnabled;
-        outStatus.backendFps = parsed.backendFps;
-        outStatus.particleCount = parsed.particleCount;
-        outStatus.solver = parsed.solver;
-        outStatus.integrator = parsed.integrator;
-        outStatus.kineticEnergy = parsed.kineticEnergy;
-        outStatus.potentialEnergy = parsed.potentialEnergy;
-        outStatus.thermalEnergy = parsed.thermalEnergy;
-        outStatus.radiatedEnergy = parsed.radiatedEnergy;
-        outStatus.totalEnergy = parsed.totalEnergy;
-        outStatus.energyDriftPct = parsed.energyDriftPct;
-        outStatus.energyEstimated = parsed.energyEstimated;
-        return response;
-    } catch (const std::exception &ex) {
-        BackendClientResponse response;
-        response.error = backendClientError("getStatus", ex.what());
-        return response;
-    } catch (...) {
-        BackendClientResponse response;
-        response.error = backendClientError("getStatus", "non-standard exception");
-        return response;
-    }
-}
-
-BackendClientResponse BackendClient::getSnapshot(std::vector<RenderParticle> &outSnapshot, std::uint32_t maxPoints)
-{
-    try {
-        maxPoints = grav_protocol::clampSnapshotPoints(maxPoints);
-        BackendClientResponse response = sendCommand(
-            std::string(grav_protocol::GetSnapshot),
-            "\"max_points\":" + std::to_string(maxPoints));
-        if (!response.ok) {
-            return response;
-        }
-
-        grav_protocol::BackendSnapshotPayload parsed{};
-        std::string parseError;
-        if (!grav_protocol::BackendJsonCodec::parseSnapshotResponse(response.raw, parsed, parseError)) {
-            response.ok = false;
-            response.error = backendClientError("getSnapshot", parseError);
-            return response;
-        }
-        outSnapshot = std::move(parsed.particles);
-        return response;
-    } catch (const std::exception &ex) {
-        BackendClientResponse response;
-        response.error = backendClientError("getSnapshot", ex.what());
-        return response;
-    } catch (...) {
-        BackendClientResponse response;
-        response.error = backendClientError("getSnapshot", "non-standard exception");
-        return response;
-    }
-}
